Renamed Predicates folder to Rules, added documentation and formatting
