+++ ChessVariantsLogic/Rules/ChessPredicates/PieceCaptured.cs
@@ -1,7 +1,8 @@
-﻿using ChessVariantsLogic;
+﻿namespace ChessVariantsLogic.Predicates;
 
-namespace ChessVariantsLogic.Predicates;
-
+/// <summary>
+/// This predicate determines if a piece has been captured when transitioning to a new board state.
+/// </summary>
 public class PieceCaptured : IPredicate
 {
     private readonly string _pieceIdentifier;
@@ -11,7 +12,11 @@ public class PieceCaptured : IPredicate
         _pieceIdentifier = pieceIdentifier;
     }
 
-
+    /// <summary>
+    /// Returns true if a piece was captured when transitioning from <paramref name="thisBoardState"/> to <paramref name="nextBoardState"/>.
+    /// </summary>
+    /// <inheritdoc/>
+    /// <returns>True if a piece was captured during transition from <paramref name="thisBoardState"/> to <paramref name="nextBoardState"/>.</returns>
     public bool Evaluate(Chessboard thisBoardState, Chessboard nextBoardState)
     {
         int amountOfPieces = Utils.FindPiecesOfType(thisBoardState, _pieceIdentifier).Count();
